--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: true
       matrix:
         os:
-          - macos-latest
+          #- macos-latest
           - ubuntu-latest
           - windows-latest
         ocaml-compiler: