--- a/src/pool/gen/dune
+++ b/src/pool/gen/dune
@@ -1,0 +1,3 @@
+(executable
+  (name mkshims))
+