cc @anuragsoni that could be of interest I suppose, on multicore, directly with domains? :)